feat(stack): move template listing to "stack templates list" and add filters
Rename `mw stack list-templates` to `mw stack templates list`, using a new
"templates" subtopic under "stack" (alias: `stack templates ls`).

Add `--category` and `--type` flags to filter the listing; both map to the
corresponding query parameters of the list-templates API and are omitted
from the request when unset.
